feat(inventory-uom-category): Refactor - UoM Category template strcuture
