+++ moduerp/inventory/static/js/products/pages/product_category_list.js
@@ -1,10 +1,9 @@
-import {get_list} from '../api/get_list.js';
+import {get_product_category_list} from '../api/get_product_category_list.js';
 
 let currentOrder = null;
 let currentDomain = null;
 let currentPage = 1;
 
 document.addEventListener("DOMContentLoaded", async () => {
-    const data = await get_list(currentDomain, currentOrder, currentPage);
-    console.log(data);
+    const data = await get_product_category_list(currentDomain, currentOrder, currentPage);
 });
