+++ moduerp/inventory/static/js/uom/utils/sort_uoms.js
@@ -1,0 +1,4 @@
+export function sortUom(data) {
+    const sortedUom = [...data].sort((a, b) => b.factor - a.factor);
+    return sortedUom
+}
