+++ moduerp/inventory/templates/products/productCategory_list.html
@@ -14,18 +14,6 @@
 {% block content %}
 <div class="flex-grow-1 overflow-auto w-100" id="product-list-view"></div>
 <script type="module" src="{% static 'js/products/pages/product_category_list.js' %}"></script>
-
-<!-- Uom Category list template -->
-<template id="uom-category-row-template">
-  <tr>
-    <td class="text-center">
-      <input type="checkbox" class="select-item form-check-input bg-dark border-indigo-5">
-    </td>
-    <td class="text-start px-3 text-light uom-category-name"></td>
-    <td class="text-start px-3 text-light uom-list"></td>
-  </tr>
-</template>
-
 <!-- Filter Model -->
 {% include "uom/_uomCategory_list_modal_filter.html"%}
 <!-- Order Model -->
