+++ moduerp/inventory/templates/uom/components/uom_category_list_toolbar.html
@@ -1,0 +1,78 @@
+{% extends "layout/toolbar.html" %}
+{% load static %}
+
+{% block toolbar_title%}
+<div name="title" class="fs-5 text-white">
+  Units of Measure Categories
+</div>
+{%endblock%}
+
+{% block toolbar_actions %} 
+<div class="d-flex flex-row align-items-center gap-3">
+  
+  <!-- Order -->
+  <div class="d-flex">
+    <!-- Toggle button -->
+    <button type="button" 
+      class="btn btn-outline-cyan-5 btn-sm rounded-start rounded-0" 
+      data-bs-toggle="button" autocomplete="off" id="order-btn">
+        <i class="bi bi-sort-down"></i> Order
+    </button>
+    <!-- Expand button -->
+    <button type="button" 
+      class="btn btn-light btn-sm rounded-end rounded-0 dropdown-toggle dropdown-toggle-split" 
+      data-bs-toggle="modal" 
+      data-bs-target="#orderModal"
+    >
+      <span class="visually-hidden">Toggle Dropdown</span>
+    </button>
+  </div>
+
+  <!-- Filter -->
+  <div class="d-flex">
+    <!-- Toggle button -->
+    <button type="button" 
+      class="btn btn-outline-cyan-5 btn-sm rounded-start rounded-0" 
+      data-bs-toggle="button" autocomplete="off" id="filter-btn">
+        <i class="bi bi-funnel"></i> Filter
+    </button>
+    <!-- Expand button -->
+    <button type="button" 
+      class="btn btn-light btn-sm rounded-end rounded-0 dropdown-toggle dropdown-toggle-split" 
+      data-bs-toggle="modal" 
+      data-bs-target="#filterModal"
+    >
+      <span class="visually-hidden">Toggle Dropdown</span>
+    </button>
+  </div>
+
+  <!-- Selected -->
+  <div class="text-cyan-5 border border-cyan-5 rounded px-3 py-1 d-inline-block small"> 
+    Selected: <span id="selected-count">0</span> 
+  </div>
+
+  <!-- Total -->
+  <div class="d-flex flex-row align-items-center gap-1">
+    <div class="fw-light"> Total:</div> 
+    <div id="toolbar-total">0</div> 
+  </div>
+
+  <!-- Navigator -->
+  <div class="d-flex flex-row align-items-center gap-1">
+    <button type="button" id="toolbar-prev-btn" class="btn btn-indigo-5 btn-sm">
+      <img src="{% static 'icons/prev.svg' %}" alt="next icon" width="10">
+    </button>
+    <button type="button" id="toolbar-next-btn" class="btn btn-indigo-5 btn-sm">
+      <img src="{% static 'icons/next.svg' %}" alt="next icon" width="10">
+    </button>      
+  </div>
+  
+  <div class="fw-light d-flex flex-row align-items-center gap-1">
+    <div id="toolbar-current-page">0</div>/<div id="toolbar-num-pages">0</div>
+  </div>
+  
+  <!-- Add New button -->
+  <a href="{% url 'inventory:uom_category_new' %}" class="btn btn-indigo-5 btn-sm">Add New</a>
+
+</div>
+{% endblock %}
