+++ moduerp/inventory/templates/uom/pages/uom_category_list.html
@@ -1,6 +1,10 @@
 {% extends "base.html" %}
 {% load static %}
 
+{% block script %} 
+<script type="module" src="{% static 'js/uom/pages/uom_category_list.js' %}"></script>
+{% endblock %}
+
 {% block title %}Inventory - Units of Measure Categories{% endblock %}
 
 {% block navbar %}
@@ -8,24 +12,11 @@
 {% endblock %}
 
 {% block toolbar %}
-{% include "uom/_uomCategory_list_toolbar.html"%}
+{% include "uom/components/uom_category_list_toolbar.html"%}
 {% endblock %}
 
 {% block content %}
-<div class="flex-grow-1 overflow-auto w-100" id="product-list-view"></div>
-<script src="{% static 'js/uom/load_uom_category_list.js' %}"></script>
-
-<!-- Uom Category list template -->
-<template id="uom-category-row-template">
-  <tr>
-    <td class="text-center">
-      <input type="checkbox" class="select-item form-check-input bg-dark border-indigo-5">
-    </td>
-    <td class="text-start px-3 text-light uom-category-name"></td>
-    <td class="text-start px-3 text-light uom-list"></td>
-  </tr>
-</template>
-
+{% include "uom/components/uom_category_list_view.html"%}
 <!-- Filter Model -->
 {% include "uom/_uomCategory_list_modal_filter.html"%}
 <!-- Order Model -->
