+++ moduerp/templates/base.html
@@ -10,6 +10,7 @@
     <!-- Bootstrap -->
     <script src="{% static 'vendor/bootstrap/bootstrap.bundle.min.js' %}"></script>
     <link href="https://cdn.jsdelivr.net/npm/bootstrap-icons/font/bootstrap-icons.css" rel="stylesheet">
+    {% block script %} {% endblock %}
 </head>
 <body class="d-flex flex-column align-items-center justify-content-center vh-100 bg-black text-white">
     <!-- Header -->
